refactor(viewer): extract shared test stubs for lib helpers
Five lib helper test suites (placement-edit, slab-edit,
linear-element-edit, metadata-clone, wall-opening-reassign) each
carried near-identical copies of the StubStoreEditor /
StubView / makeStore harness — roughly 50-80 lines per file with
small drift. Centralise them in apps/viewer/src/lib/__test__/stubs.ts
(OverlayEntity, StubStoreEditor, StubView, makeStubDataStore) so the
contract evolves in one place and the test files focus on fixtures
and assertions.

Net 195 lines removed across the 5 suites; all 49 tests still pass.
